split dtype
