+++ pandas/core/arrays/sparse/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa: F401
 
 from .accessor import SparseAccessor, SparseFrameAccessor
-from .sparse import BlockIndex, IntIndex, SparseArray, SparseDtype, _make_index
+from .dtype import SparseDtype
+from .sparse import BlockIndex, IntIndex, SparseArray, _make_index
